+++ dojos/effect-ts/CHANGELOG.md
@@ -1,11 +1,11 @@
-# @dojocho/effect-ts
+# @dojofoo/effect-ts
 
 ## 0.0.5
 
 ### Patch Changes
 
 - Updated dependencies [d87d2e2]
-  - @dojocho/config@0.0.5
+  - @dojofoo/config@0.0.5
 
 ## 0.0.4